# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GURU HOST API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /deployments:
    get:
      operationId: getDeployments
      tags: [deployments]
      summary: Get all deployments for current user
      responses:
        "200":
          description: List of deployments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deployment"
    post:
      operationId: createDeployment
      tags: [deployments]
      summary: Create a new bot deployment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeploymentRequest"
      responses:
        "200":
          description: Deployment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
        "402":
          description: Insufficient credits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /deployments/{id}:
    get:
      operationId: getDeployment
      tags: [deployments]
      summary: Get a deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deployment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
    delete:
      operationId: deleteDeployment
      tags: [deployments]
      summary: Delete a deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /deployments/{id}/restart:
    post:
      operationId: restartDeployment
      tags: [deployments]
      summary: Restart a deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Restarted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /deployments/{id}/pause:
    post:
      operationId: pauseDeployment
      tags: [deployments]
      summary: Pause/stop a deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /deployments/{id}/resume:
    post:
      operationId: resumeDeployment
      tags: [deployments]
      summary: Resume a paused deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /deployments/{id}/logs:
    get:
      operationId: getDeploymentLogs
      tags: [deployments]
      summary: Get logs for a deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: lines
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogsResponse"
  /deployments/{id}/env:
    get:
      operationId: getDeploymentEnv
      tags: [deployments]
      summary: Get env vars for a deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Env vars
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EnvVar"
    post:
      operationId: setDeploymentEnv
      tags: [deployments]
      summary: Set env vars for a deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                vars:
                  type: array
                  items:
                    $ref: "#/components/schemas/EnvVar"
              required: [vars]
      responses:
        "200":
          description: Env vars updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/bots:
    post:
      operationId: adminCreateBot
      tags: [admin]
      summary: Admin - add a new bot to catalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBotRequest"
      responses:
        "200":
          description: Bot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Deployment:
      type: object
      properties:
        id:
          type: string
        userId:
          type: number
        botId:
          type: number
        appName:
          type: string
        repoUrl:
          type: string
        sessionId:
          type: string
        status:
          type: string
          enum: [building, running, paused, crashed, deleted]
        herokuAppId:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        bot:
          $ref: "#/components/schemas/Bot"
      required: [id, userId, botId, appName, repoUrl, sessionId, status, createdAt, updatedAt]
    CreateDeploymentRequest:
      type: object
      properties:
        botId:
          type: number
        repoUrl:
          type: string
        sessionId:
          type: string
        envVars:
          type: array
          items:
            $ref: "#/components/schemas/EnvVar"
      required: [botId, repoUrl, sessionId]
    EnvVar:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
      required: [key, value]
    LogsResponse:
      type: object
      properties:
        logs:
          type: string
        lines:
          type: array
          items:
            type: string
      required: [logs, lines]
    Bot:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        description:
          type: string
        repoUrl:
          type: string
        imageUrl:
          type: string
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, description, repoUrl, isActive, createdAt]
    CreateBotRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        repoUrl:
          type: string
        imageUrl:
          type: string
        features:
          type: array
          items:
            type: string
      required: [name, description, repoUrl]
...